refactor: 애니 상세페이지 파일 구조 #93
